+++ crates/platform-linux/src/ffi/libc_surface.rs
@@ -5,19 +5,20 @@
 //! replacement inventory honest: this file *is* the checklist.
 
 pub use libc::{
-    c_char, c_int, close, dirent, faccessat, fchmodat, fdopendir, fstatat, fsync, kill, mkdirat,
-    nfds_t, openat, pid_t, pipe2, poll, pollfd, posix_spawn, posix_spawn_file_actions_addchdir_np,
-    posix_spawn_file_actions_adddup2, posix_spawn_file_actions_addopen,
-    posix_spawn_file_actions_destroy, posix_spawn_file_actions_init, posix_spawn_file_actions_t,
-    posix_spawnattr_destroy, posix_spawnattr_init, posix_spawnattr_setflags,
-    posix_spawnattr_setpgroup, posix_spawnattr_t, read, readdir, readlinkat, sighandler_t, signal,
-    stat, symlinkat, syscall, tcgetattr, tcsetattr, termios, unlinkat, waitpid, winsize, write,
-    SYS_pidfd_open, SYS_renameat2, AT_FDCWD, AT_REMOVEDIR, AT_SYMLINK_NOFOLLOW, DIR, DT_DIR,
-    DT_LNK, DT_REG, O_APPEND, O_CLOEXEC, O_CREAT, O_DIRECTORY, O_EXCL, O_RDONLY, O_RDWR, O_TRUNC,
-    O_WRONLY, POLLIN, POSIX_SPAWN_SETPGROUP, RENAME_NOREPLACE, R_OK, SIGHUP, SIGINT, SIGKILL,
-    SIGTERM, SIG_ERR, STDERR_FILENO, STDIN_FILENO, STDOUT_FILENO, S_IFDIR, S_IFLNK, S_IFMT,
-    S_IFREG, S_ISGID, S_ISUID, S_ISVTX, TCSADRAIN, TIOCGWINSZ, WEXITSTATUS, WIFEXITED, WIFSIGNALED,
-    WNOHANG, WTERMSIG, W_OK, X_OK,
+    c_char, c_int, c_short, close, dirent, faccessat, fchmodat, fdopendir, fstatat, fsync, kill,
+    mkdirat, nfds_t, openat, pid_t, pipe2, poll, pollfd, posix_spawn,
+    posix_spawn_file_actions_addchdir_np, posix_spawn_file_actions_adddup2,
+    posix_spawn_file_actions_addopen, posix_spawn_file_actions_destroy,
+    posix_spawn_file_actions_init, posix_spawn_file_actions_t, posix_spawnattr_destroy,
+    posix_spawnattr_init, posix_spawnattr_setflags, posix_spawnattr_setpgroup, posix_spawnattr_t,
+    read, readdir, readlinkat, sighandler_t, signal, stat, symlinkat, syscall, tcgetattr,
+    tcsetattr, termios, unlinkat, waitpid, winsize, write, SYS_pidfd_open, SYS_renameat2, AT_FDCWD,
+    AT_REMOVEDIR, AT_SYMLINK_NOFOLLOW, DIR, DT_DIR, DT_LNK, DT_REG, ESRCH, O_APPEND, O_CLOEXEC,
+    O_CREAT, O_DIRECTORY, O_EXCL, O_RDONLY, O_RDWR, O_TRUNC, O_WRONLY, POLLIN,
+    POSIX_SPAWN_SETPGROUP, RENAME_NOREPLACE, R_OK, SIGHUP, SIGINT, SIGKILL, SIGTERM, SIG_ERR,
+    STDERR_FILENO, STDIN_FILENO, STDOUT_FILENO, S_IFDIR, S_IFLNK, S_IFMT, S_IFREG, S_ISGID,
+    S_ISUID, S_ISVTX, TCSADRAIN, TIOCGWINSZ, WEXITSTATUS, WIFEXITED, WIFSIGNALED, WNOHANG,
+    WTERMSIG, W_OK, X_OK,
 };
 
 // R2/R3 containment (Rusty-Mill fs slice, `sys::openat2`): `openat2(2)`